move fragments to other namespace
